fn s3_request: additional closure to get the actual size of request
- Earlier we only had "assumed" size (which was always correct).
But if the actual size depends more on the response, we need a way to
get that size based on the response, for the purpose of fine-tuning the
timeouts.

- Additionally, s/u64/usize in most places
- Initial version of a `process` function on ListObjects

+++ src/config.rs
@@ -73,7 +73,7 @@ pub struct SpecificTimings {
     /// The initial estimate of extra timeout per unit (byte or object)
     pub seconds_per_unit: f64,
     /// The amount of units in a request, below which it does not affect estimation
-    pub minimum_units_for_estimation: u64,
+    pub minimum_units_for_estimation: usize,
 }
 
 impl SpecificTimings {
